use new config and make the components linux compatible
